Refactor to_chars_float implementation details
