+++ include/boost/charconv/detail/to_chars_integer_impl.hpp
@@ -12,6 +12,7 @@
 #include <boost/charconv/detail/to_chars_result.hpp>
 #include <boost/charconv/detail/integer_search_trees.hpp>
 #include <boost/charconv/detail/emulated128.hpp>
+#include <boost/charconv/detail/apply_sign.hpp>
 #include <limits>
 #include <system_error>
 #include <type_traits>
